ncm: Moved main()
- Moved non-public APIs also
